Rename es indexer as fileCentricIndexer
