+++ src/jobs/processAnalysisEvent.ts
@@ -20,7 +20,7 @@ import AnalysisUpdateEvent from '../external/kafka/messages/AnalysisUpdateEvent'
 import { convertAnalysesToFileDocuments } from '../external/analysisConverter';
 import { saveAndIndexFilesFromRdpcData } from '../services/fileManager';
 import { isRestricted } from '../services/utils/fileUtils';
-import { getIndexer } from '../services/indexer';
+import { getFileCentricIndexer } from '../services/fileCentricIndexer';
 import * as fileService from '../data/files';
 import PromisePool from '@supercharge/promise-pool';
 
@@ -28,10 +28,10 @@ import Logger from '../logger';
 const logger = Logger('Job:ProcessAnalysisEvent');
 
 async function handleSongPublishedAnalysis(analysis: any, dataCenterId: string) {
-  const rdpcFileDocuments = await convertAnalysesToFileDocuments([analysis], dataCenterId);
-  const indexer = await getIndexer();
-  await saveAndIndexFilesFromRdpcData(rdpcFileDocuments, dataCenterId, indexer);
-  await indexer.release();
+	const rdpcFileDocuments = await convertAnalysesToFileDocuments([analysis], dataCenterId);
+	const indexer = await getFileCentricIndexer();
+	await saveAndIndexFilesFromRdpcData(rdpcFileDocuments, dataCenterId, indexer);
+	await indexer.release();
 }
 
 /**
@@ -42,31 +42,31 @@ async function handleSongPublishedAnalysis(analysis: any, dataCenterId: string) 
  * @param status
  */
 async function handleSongUnpublishedAnalysis(analysisId: string, status: string): Promise<void> {
-  // Get files based on analysis ID
-  // TODO: This should create the file if we don't have it yet
-  const files = await fileService.getFilesByAnalysisId(analysisId);
-  if (files.length === 0) {
-    logger.info(`No stored files for analysis ${analysisId}. No processing to do.`);
-    return;
-  }
+	// Get files based on analysis ID
+	// TODO: This should create the file if we don't have it yet
+	const files = await fileService.getFilesByAnalysisId(analysisId);
+	if (files.length === 0) {
+		logger.info(`No stored files for analysis ${analysisId}. No processing to do.`);
+		return;
+	}
 
-  logger.info(`Updating Song status to ${status} for files ${files.map(file => file.objectId)}`);
-  await PromisePool.withConcurrency(10)
-    .for(files)
-    .handleError((error, file) => {
-      logger.error(`Failure updating file ${file.objectId} status to ${status}: ${error}`);
-    })
-    .process(async file => {
-      await fileService.updateFileSongPublishStatus(file.objectId, { status });
-    });
+	logger.info(`Updating Song status to ${status} for files ${files.map(file => file.objectId)}`);
+	await PromisePool.withConcurrency(10)
+		.for(files)
+		.handleError((error, file) => {
+			logger.error(`Failure updating file ${file.objectId} status to ${status}: ${error}`);
+		})
+		.process(async file => {
+			await fileService.updateFileSongPublishStatus(file.objectId, { status });
+		});
 
-  // Remove these files from the restricted indices
-  const restrictedFiles = files.filter(isRestricted);
-  if (restrictedFiles.length) {
-    const indexer = await getIndexer();
-    await indexer.removeFilesFromRestricted(files);
-    await indexer.release();
-  }
+	// Remove these files from the restricted indices
+	const restrictedFiles = files.filter(isRestricted);
+	if (restrictedFiles.length) {
+		const indexer = await getFileCentricIndexer();
+		await indexer.removeFilesFromRestricted(files);
+		await indexer.release();
+	}
 }
 
 /**
@@ -74,25 +74,25 @@ async function handleSongUnpublishedAnalysis(analysisId: string, status: string)
  * @param analysisEvent
  */
 const processAnalysisEvent = async (analysisEvent: AnalysisUpdateEvent): Promise<void> => {
-  const { analysis, analysisId, state, songServerId } = analysisEvent;
+	const { analysis, analysisId, state, songServerId } = analysisEvent;
 
-  try {
-    logger.info(
-      `START - processing song analysis event from data-center ${songServerId} for analysisId ${analysisId} with state ${state}`,
-    );
+	try {
+		logger.info(
+			`START - processing song analysis event from data-center ${songServerId} for analysisId ${analysisId} with state ${state}`,
+		);
 
-    if (state === 'PUBLISHED') {
-      await handleSongPublishedAnalysis(analysis, songServerId);
-    } else {
-      // Unpublish or Suppress
-      await handleSongUnpublishedAnalysis(analysisId, state);
-    }
+		if (state === 'PUBLISHED') {
+			await handleSongPublishedAnalysis(analysis, songServerId);
+		} else {
+			// Unpublish or Suppress
+			await handleSongUnpublishedAnalysis(analysisId, state);
+		}
 
-    logger.info(
-      `DONE - processing song analysis event from data-center ${songServerId} for analysisId ${analysis.analysisId}`,
-    );
-  } catch (e) {
-    logger.error(`FAILURE - processing analysis event failed`, e);
-  }
+		logger.info(
+			`DONE - processing song analysis event from data-center ${songServerId} for analysisId ${analysis.analysisId}`,
+		);
+	} catch (e) {
+		logger.error(`FAILURE - processing analysis event failed`, e);
+	}
 };
 export default processAnalysisEvent;
