+++ src/routers/admin.ts
@@ -24,278 +24,278 @@ import * as fileService from '../data/files';
 import reindexDataCenter from '../jobs/reindexDataCenter';
 import Logger, { unknownToString } from '../logger';
 import { updateFileFromExternalSources } from '../services/fileManager';
-import { getIndexer } from '../services/indexer';
+import { getFileCentricIndexer } from '../services/fileCentricIndexer';
 import wrapAsync from '../utils/wrapAsync';
 import validator from './common/validator';
 
 const logger = Logger('Admin.Router');
 
 type UpdateFileSummary = {
-  total: number;
-  ids: string[];
+	total: number;
+	ids: string[];
 };
 function fileSummaryResponse(files: fileService.File[]): UpdateFileSummary {
-  const total = files.length;
-  const ids = files.map(file => file.objectId);
-  return { total, ids };
+	const total = files.length;
+	const ids = files.map(file => file.objectId);
+	return { total, ids };
 }
 
 const createAdminRouter = (config: AppConfig, authFilter: (scopes: string[]) => RequestHandler) => {
-  const router = Router();
+	const router = Router();
 
-  // Ensure all routes in this router are restricted to those with WRITE scope only.
-  router.use(authFilter([config.auth.writeScope]));
+	// Ensure all routes in this router are restricted to those with WRITE scope only.
+	router.use(authFilter([config.auth.writeScope]));
 
-  /**
-   * /index/:datacenter
-   * Request to re-index all analyses from a datacenter.
-   */
-  router.post(
-    '/index/:datacenter',
-    wrapAsync(async (req: Request, res: Response) => {
-      // const dataCenterId = req.params.datacenter;
-      // TODO: Current config hardcodes a single data center instead of retrieving connection details from data center registry
+	/**
+	 * /index/:datacenter
+	 * Request to re-index all analyses from a datacenter.
+	 */
+	router.post(
+		'/index/:datacenter',
+		wrapAsync(async (req: Request, res: Response) => {
+			// const dataCenterId = req.params.datacenter;
+			// TODO: Current config hardcodes a single data center instead of retrieving connection details from data center registry
 
-      // Swagger UI isn't sending single studies as an array, so we need to parse it
-      const studies: string[] = ((typeof req.query.study === 'string' ? [req.query.study] : req.query.study) ||
-        []) as string[];
+			// Swagger UI isn't sending single studies as an array, so we need to parse it
+			const studies: string[] = ((typeof req.query.study === 'string' ? [req.query.study] : req.query.study) ||
+				[]) as string[];
 
-      const dataCenterId = config.datacenter.dataCenterId;
-      reindexDataCenter(dataCenterId, studies);
-      res.status(200).json({ status: 'OK', message: `request submitted to index data center: ${dataCenterId}` });
-      return;
-    }),
-  );
+			const dataCenterId = config.datacenter.dataCenterId;
+			reindexDataCenter(dataCenterId, studies);
+			res.status(200).json({ status: 'OK', message: `request submitted to index data center: ${dataCenterId}` });
+			return;
+		}),
+	);
 
-  /**
-   * '/promote/:stage'
-   * Add admin promote value to files based on a filter
-   */
-  router.post(
-    '/promote/:stage',
-    wrapAsync(async (req: Request, res: Response) => {
-      // Get Params:
-      try {
-        const filter = validator.fileFilter(req.body);
-        const stage = validator.embargoStage(req.params.stage);
+	/**
+	 * '/promote/:stage'
+	 * Add admin promote value to files based on a filter
+	 */
+	router.post(
+		'/promote/:stage',
+		wrapAsync(async (req: Request, res: Response) => {
+			// Get Params:
+			try {
+				const filter = validator.fileFilter(req.body);
+				const stage = validator.embargoStage(req.params.stage);
 
-        const dryRun = req.query.dryRun === 'true';
+				const dryRun = req.query.dryRun === 'true';
 
-        if (dryRun) {
-          res.status(200).json(await dryRunFileFilter(filter));
-          return;
-        }
+				if (dryRun) {
+					res.status(200).json(await dryRunFileFilter(filter));
+					return;
+				}
 
-        try {
-          // Update files in DB
-          const updatedFiles = await fileService.adminPromote(filter, stage, {
-            returnDocuments: true,
-          });
+				try {
+					// Update files in DB
+					const updatedFiles = await fileService.adminPromote(filter, stage, {
+						returnDocuments: true,
+					});
 
-          if (!updatedFiles) {
-            res.status(400).json({ status: 'Failure', message: `No files updated.` });
-            return;
-          }
+					if (!updatedFiles) {
+						res.status(400).json({ status: 'Failure', message: `No files updated.` });
+						return;
+					}
 
-          const result = await indexUpdatedFiles(updatedFiles);
+					const result = await indexUpdatedFiles(updatedFiles);
 
-          const response = {
-            message: `Successfully updated ${result.fileSummary.total} file(s). adminPromote value set to ${stage}`,
-            ...result.fileSummary,
-            errors: result.errors,
-          };
-          res.status(200).json(response);
-          return;
-        } catch (e) {
-          res.status(500).json({ status: 'Error', message: `Unexpected error updating files: ${e}` });
-          return;
-        }
-      } catch (error) {
-        // Catch Param Validation Errors
-        res.status(400).json({ error: unknownToString(error) });
-        return;
-      }
-    }),
-  );
+					const response = {
+						message: `Successfully updated ${result.fileSummary.total} file(s). adminPromote value set to ${stage}`,
+						...result.fileSummary,
+						errors: result.errors,
+					};
+					res.status(200).json(response);
+					return;
+				} catch (e) {
+					res.status(500).json({ status: 'Error', message: `Unexpected error updating files: ${e}` });
+					return;
+				}
+			} catch (error) {
+				// Catch Param Validation Errors
+				res.status(400).json({ error: unknownToString(error) });
+				return;
+			}
+		}),
+	);
 
-  /**
-   * '/demote/:stage'
-   * Add admin restrict value to files based on a filter
-   */
-  router.post(
-    '/demote/:stage',
-    wrapAsync(async (req: Request, res: Response) => {
-      // Get Params:
-      try {
-        const filter = validator.fileFilter(req.body);
-        const stage = validator.embargoStage(req.params.stage);
+	/**
+	 * '/demote/:stage'
+	 * Add admin restrict value to files based on a filter
+	 */
+	router.post(
+		'/demote/:stage',
+		wrapAsync(async (req: Request, res: Response) => {
+			// Get Params:
+			try {
+				const filter = validator.fileFilter(req.body);
+				const stage = validator.embargoStage(req.params.stage);
 
-        const dryRun = req.query.dryRun === 'true';
+				const dryRun = req.query.dryRun === 'true';
 
-        if (dryRun) {
-          res.status(200).json(await dryRunFileFilter(filter));
-          return;
-        }
+				if (dryRun) {
+					res.status(200).json(await dryRunFileFilter(filter));
+					return;
+				}
 
-        try {
-          // Update files in DB
-          const updatedFiles = await fileService.adminDemote(filter, stage, {
-            returnDocuments: true,
-          });
+				try {
+					// Update files in DB
+					const updatedFiles = await fileService.adminDemote(filter, stage, {
+						returnDocuments: true,
+					});
 
-          if (!updatedFiles) {
-            res.status(400).send(`No files updated.`);
-            return;
-          }
+					if (!updatedFiles) {
+						res.status(400).send(`No files updated.`);
+						return;
+					}
 
-          const result = await indexUpdatedFiles(updatedFiles);
+					const result = await indexUpdatedFiles(updatedFiles);
 
-          const response = {
-            message: `Successfully updated ${result.fileSummary.total} file(s). adminDemote value set to ${stage}`,
-            ...result.fileSummary,
-            errors: result.errors,
-          };
-          res.status(200).json(response);
-          return;
-        } catch (e) {
-          res.status(500).send(`Unexpected error updating files: ${e}`);
-          return;
-        }
-      } catch (error) {
-        // Catch Param Validation Errors
-        res.status(400).send({ error: unknownToString(error) });
-        return;
-      }
-    }),
-  );
+					const response = {
+						message: `Successfully updated ${result.fileSummary.total} file(s). adminDemote value set to ${stage}`,
+						...result.fileSummary,
+						errors: result.errors,
+					};
+					res.status(200).json(response);
+					return;
+				} catch (e) {
+					res.status(500).send(`Unexpected error updating files: ${e}`);
+					return;
+				}
+			} catch (error) {
+				// Catch Param Validation Errors
+				res.status(400).send({ error: unknownToString(error) });
+				return;
+			}
+		}),
+	);
 
-  /**
-   * '/clinicalExemption/remove'
-   * Remove clinical exemption from files
-   */
-  router.post(
-    '/clinicalExemption/remove',
-    wrapAsync(async (req: Request, res: Response) => {
-      // Get Params:
-      try {
-        const filter = validator.fileFilter(req.body);
+	/**
+	 * '/clinicalExemption/remove'
+	 * Remove clinical exemption from files
+	 */
+	router.post(
+		'/clinicalExemption/remove',
+		wrapAsync(async (req: Request, res: Response) => {
+			// Get Params:
+			try {
+				const filter = validator.fileFilter(req.body);
 
-        const dryRun = req.query.dryRun === 'true';
+				const dryRun = req.query.dryRun === 'true';
 
-        if (dryRun) {
-          res.status(200).json(await dryRunFileFilter(filter));
-          return;
-        }
+				if (dryRun) {
+					res.status(200).json(await dryRunFileFilter(filter));
+					return;
+				}
 
-        try {
-          // Update files in DB
-          const updatedFiles = await fileService.removeClinicalExemption(filter, {
-            returnDocuments: true,
-          });
+				try {
+					// Update files in DB
+					const updatedFiles = await fileService.removeClinicalExemption(filter, {
+						returnDocuments: true,
+					});
 
-          if (!updatedFiles) {
-            res.status(400).send(`No files updated.`);
-            return;
-          }
+					if (!updatedFiles) {
+						res.status(400).send(`No files updated.`);
+						return;
+					}
 
-          const response = {
-            message: `Successfully updated ${updatedFiles.length} file(s), removing any clinical exemptions.`,
-          };
-          res.status(200).json(response);
-          return;
-        } catch (e) {
-          res.status(500).send(`Unexpected error updating files: ${e}`);
-          return;
-        }
-      } catch (error) {
-        // Catch Param Validation Errors
-        res.status(400).send({ error: unknownToString(error) });
-        return;
-      }
-    }),
-  );
+					const response = {
+						message: `Successfully updated ${updatedFiles.length} file(s), removing any clinical exemptions.`,
+					};
+					res.status(200).json(response);
+					return;
+				} catch (e) {
+					res.status(500).send(`Unexpected error updating files: ${e}`);
+					return;
+				}
+			} catch (error) {
+				// Catch Param Validation Errors
+				res.status(400).send({ error: unknownToString(error) });
+				return;
+			}
+		}),
+	);
 
-  /**
-   * '/clinicalExemption/:reason'
-   * Apply a clinical exemption to a selection of files
-   * Files with a clinical exemption do not require core complete clincial data to be released to the index
-   */
-  router.post(
-    '/clinicalExemption/:reason',
-    wrapAsync(async (req: Request, res: Response) => {
-      // Get Params:
-      try {
-        const filter = validator.fileFilter(req.body);
-        const reason = validator.clinicalExemption(req.params.reason);
+	/**
+	 * '/clinicalExemption/:reason'
+	 * Apply a clinical exemption to a selection of files
+	 * Files with a clinical exemption do not require core complete clincial data to be released to the index
+	 */
+	router.post(
+		'/clinicalExemption/:reason',
+		wrapAsync(async (req: Request, res: Response) => {
+			// Get Params:
+			try {
+				const filter = validator.fileFilter(req.body);
+				const reason = validator.clinicalExemption(req.params.reason);
 
-        const dryRun = req.query.dryRun === 'true';
+				const dryRun = req.query.dryRun === 'true';
 
-        if (dryRun) {
-          res.status(200).json(await dryRunFileFilter(filter));
-          return;
-        }
+				if (dryRun) {
+					res.status(200).json(await dryRunFileFilter(filter));
+					return;
+				}
 
-        try {
-          // Update files in DB
-          const updatedFiles = await fileService.applyClinicalExemption(filter, reason, {
-            returnDocuments: true,
-          });
+				try {
+					// Update files in DB
+					const updatedFiles = await fileService.applyClinicalExemption(filter, reason, {
+						returnDocuments: true,
+					});
 
-          if (!updatedFiles) {
-            res.status(400).send(`No files updated.`);
-            return;
-          }
+					if (!updatedFiles) {
+						res.status(400).send(`No files updated.`);
+						return;
+					}
 
-          const response = {
-            message: `Successfully updated ${updatedFiles.length} file(s) with clinical exemption: ${reason}.`,
-          };
-          res.status(200).json(response);
-          return;
-        } catch (e) {
-          res.status(500).send(`Unexpected error updating files: ${e}`);
-          return;
-        }
-      } catch (error) {
-        // Catch Param Validation Errors
-        res.status(400).send({ error: unknownToString(error) });
-        return;
-      }
-    }),
-  );
+					const response = {
+						message: `Successfully updated ${updatedFiles.length} file(s) with clinical exemption: ${reason}.`,
+					};
+					res.status(200).json(response);
+					return;
+				} catch (e) {
+					res.status(500).send(`Unexpected error updating files: ${e}`);
+					return;
+				}
+			} catch (error) {
+				// Catch Param Validation Errors
+				res.status(400).send({ error: unknownToString(error) });
+				return;
+			}
+		}),
+	);
 
-  return router;
+	return router;
 };
 
 async function dryRunFileFilter(filter: fileService.FileFilter) {
-  const selectedFiles = await fileService.getFiles(filter);
-  const fileSummary = fileSummaryResponse(selectedFiles);
+	const selectedFiles = await fileService.getFiles(filter);
+	const fileSummary = fileSummaryResponse(selectedFiles);
 
-  return { message: 'DRY RUN ONLY - No changes made.', ...fileSummary };
+	return { message: 'DRY RUN ONLY - No changes made.', ...fileSummary };
 }
 
 async function indexUpdatedFiles(updatedFiles: fileService.File[]) {
-  const errors: Record<string, Error> = {};
-  const indexer = await getIndexer();
-  // Update file indices with changes
-  const { results } = await PromisePool.withConcurrency(20)
-    .for(updatedFiles)
-    .handleError((e, file) => {
-      logger.error(`Update Doc Error: ${e}`);
-      logger.error(`Update Doc Error: ${e.stack}`);
-      errors[file.objectId] = e;
-    })
-    .process(async file => {
-      logger.debug(`Updating and reindexing file: ${file.fileId}`);
-      const recalculatedFile = await updateFileFromExternalSources(file);
-      await indexer.updateRestrictedFile(recalculatedFile);
-      return file;
-    });
+	const errors: Record<string, Error> = {};
+	const indexer = await getFileCentricIndexer();
+	// Update file indices with changes
+	const { results } = await PromisePool.withConcurrency(20)
+		.for(updatedFiles)
+		.handleError((e, file) => {
+			logger.error(`Update Doc Error: ${e}`);
+			logger.error(`Update Doc Error: ${e.stack}`);
+			errors[file.objectId] = e;
+		})
+		.process(async file => {
+			logger.debug(`Updating and reindexing file: ${file.fileId}`);
+			const recalculatedFile = await updateFileFromExternalSources(file);
+			await indexer.updateRestrictedFile(recalculatedFile);
+			return file;
+		});
 
-  await indexer.release();
+	await indexer.release();
 
-  const fileSummary = fileSummaryResponse(results);
+	const fileSummary = fileSummaryResponse(results);
 
-  return { fileSummary, errors };
+	return { fileSummary, errors };
 }
 
 export default createAdminRouter;
